project_backend/requirements.txt
-i https://pypi.org/simple
aniso8601==10.0.1
attrs==26.1.0; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
click==8.3.2; python_version >= '3.10'
flask==3.1.3; python_version >= '3.9'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
flask-restx==1.3.2; python_version >= '3.9'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.4.0; python_version >= '3.10'
importlib-resources==7.1.0; python_version >= '3.10'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.26.0; python_version >= '3.10'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
markupsafe==3.0.3; python_version >= '3.9'
pyjwt==2.12.1; python_version >= '3.9'
python-decouple==3.8
referencing==0.37.0; python_version >= '3.10'
rpds-py==0.30.0; python_version >= '3.10'
sqlalchemy==2.0.49; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
werkzeug==3.1.8; python_version >= '3.9'
